+++ bindings/angular1/directives/splitterContent.js
@@ -3,6 +3,15 @@
  */
 
 /**
+ * @attribute var
+ * @initonly
+ * @type {String}
+ * @description
+ *   [en]Variable name to refer this splitter content.[/en]
+ *   [ja]このスプリッターコンポーネントを参照するための名前を指定します。[/ja]
+ */
+
+/**
  * @attribute ons-destroy
  * @initonly
  * @type {Expression}