+++ bindings/vue/src/components/VOnsNavigator.vue
@@ -8,7 +8,6 @@
 
 <script>
   import 'onsenui/esm/elements/ons-navigator';
-  import ons from 'onsenui/esm';
   import { hasOptions, selfProvider, deriveEvents, deriveDBB } from '../mixins';
 
   export default {
@@ -44,7 +43,7 @@
       },
       _findScrollPage(page) {
         const nextPage = page._contentElement.children.length === 1
-          && ons._util.getTopPage(page._contentElement.children[0]);
+          && this.$ons._ons._util.getTopPage(page._contentElement.children[0]);
         return nextPage ? this._findScrollPage(nextPage) : page;
       },
       _setPagesVisibility(start, end, visibility) {