+++ bindings/vue/src/setup.js
@@ -1,27 +1,23 @@
-import ons from 'onsenui/esm';
-
-const $ons = Object.keys(ons)
-  .filter(k => [
-    /^is/,
-    /^disable/,
-    /^enable/,
-    /^mock/,
-    /^open/,
-    /^set/,
-    /animit/,
-    /elements/,
-    /fastClick/,
-    /GestureDetector/,
-    /notification/,
-    /orientation/,
-    /platform/,
-    /ready/,
-  ].some(t => k.match(t)))
-  .reduce((r, k) => {
-    r[k] = ons[k];
-    return r;
-  }, { _ons: ons });
-
-export {
-  $ons,
-};
+export default function(ons) {
+  return Object.keys(ons)
+    .filter(k => [
+      /^is/,
+      /^disable/,
+      /^enable/,
+      /^mock/,
+      /^open/,
+      /^set/,
+      /animit/,
+      /elements/,
+      /fastClick/,
+      /GestureDetector/,
+      /notification/,
+      /orientation/,
+      /platform/,
+      /ready/,
+    ].some(t => k.match(t)))
+    .reduce((r, k) => {
+      r[k] = ons[k];
+      return r;
+    }, { _ons: ons });
+}