+++ core/src/elements/ons-page.js
@@ -452,6 +452,7 @@ export default class PageElement extends BaseElement {
     this.constructor.events.forEach(event => {
       const key = 'on' + event.charAt(0).toUpperCase() + event.slice(1);
       Object.defineProperty(this, key, {
+        configurable: true,
         enumerable: true,
         get: () => this[`_${key}`],
         set: value => {