+++ core/src/index.esm.js
@@ -1,5 +1,5 @@
 import ons from './ons'; // External dependency, always hoisted
-import setup from './setup';
+import setup from './setup'; // Add polyfills
 
 setup(ons); // Setup initial listeners